refactor: move htmlRenderConfig and linkMarkOption to a new config.go file for better organization and maintainability
